requirements.txt
alpaca-py>=0.21.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
pyyaml>=6.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pytz>=2023.3
pytest>=7.0.0
flask>=3.0.0
